src/crawler/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.6.20
    # via requests
chardet==3.0.4
    # via requests
idna==2.10
    # via requests
numpy==1.19.2
    # via pandas
pandas==1.1.3
    # via -r requirements.in
plotly==4.14.3
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
python-dateutil==2.8.1
    # via pandas
pytz==2020.1
    # via pandas
requests==2.24.0
    # via -r requirements.in
retrying==1.3.3
    # via plotly
schedule==1.1.0
    # via -r requirements.in
six==1.15.0
    # via
    #   plotly
    #   python-dateutil
    #   retrying
sqlalchemy==1.3.23
    # via -r requirements.in
urllib3==1.25.10
    # via requests
